read from input stream in interpreter, move input parsing to PrimitiveType, move all exception classes to exception.py
