+++ tests/builtins.p
@@ -1,7 +1,7 @@
 OUTPUT "LEGNTH(ABCDEFG): " & LENGTH("ABCDEFG")
 OUTPUT "SUBSTRING(ABCDEFG, 4, 2): " & SUBSTRING("ABCDEFG", 4, 2)
-OUTPUT "UCASE(ThiS 1s a T3st 112!): " & UCASE("ThiS 1s a T3st!")
-OUTPUT "LCASE(ThiS 1s a T3st 112!): " & LCASE("ThiS 1s a T3st!")
+OUTPUT "UCASE(ThiS 1s a T3st!): " & UCASE("ThiS 1s a T3st!")
+OUTPUT "LCASE(ThiS 1s a T3st!): " & LCASE("ThiS 1s a T3st!")
 OUTPUT "RANDOM():" & RANDOM()
 OUTPUT "DIV(10, 3): " & DIV(10, 3)
 OUTPUT "MOD(10, 3): " & MOD(10, 3)
